Starting the refactoring of account_cost_spread. Added domain for spread_account_id to make the module more multi-company compliant.
