+++ account_cost_spread/views/account_invoice_line_view.xml
@@ -2,7 +2,7 @@
 <openerp>
     <data>
 
-        <record model="ir.ui.view" id="view_invoice_line_spread">
+        <record id="view_invoice_line_spread" model="ir.ui.view">
             <field name="name">account.invoice.line.spread</field>
             <field name="model">account.invoice.line</field>
             <field name="arch" type="xml">
@@ -15,7 +15,8 @@
                                 <field name="price_subtotal" readonly="1"/>
                             </group>
                             <group>
-                                <field name="spread_account_id" required="1"/>
+                                <field name="company_id" invisible="1"/>
+                                <field name="spread_account_id" required="1" domain="[('company_id', '=', company_id), ('type', '&lt;&gt;', 'view')]"/>
                                 <field name="period_number" required="1"/>
                                 <field name="period_type" required="1"/>
                                 <field name="spread_date"/>
@@ -39,33 +40,6 @@
                         </field>
                     </sheet>
                 </form>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_supplier_invoice_spread">
-            <field name="name">account.invoice.supplier.form</field>
-            <field name="model">account.invoice</field>
-            <field name="inherit_id" ref="account.invoice_supplier_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='invoice_line']/tree/field[@name='quantity']"
-                       position="before">
-                    <button name="spread_details" string="Spread" icon="terp-gtk-jump-to-ltr"
-                            type="object"/>
-                </xpath>
-            </field>
-        </record>
-
-        <record model="ir.ui.view" id="view_customer_invoice_spread">
-            <field name="name">account.invoice.customer.form</field>
-            <field name="model">account.invoice</field>
-            <field name="inherit_id" ref="account.invoice_form"/>
-            <field name="arch" type="xml">
-                <xpath expr="//field[@name='invoice_line']/tree/field[@name='quantity']"
-                       position="before">
-                    <button name="spread_details" string="Spread" icon="terp-gtk-jump-to-ltr"
-                            type="object"/>
-                </xpath>
-
             </field>
         </record>
