+++ account_cost_spread/views/res_config_view.xml
@@ -2,7 +2,7 @@
 <openerp>
     <data>
 
-        <record model="ir.ui.view" id="view_account_config_settings">
+        <record id="view_account_config_settings" model="ir.ui.view">
             <field name="name">account.config.settings.inherit</field>
             <field name="inherit_id" ref="account.view_account_config_settings"/>
             <field name="model">account.config.settings</field>
